src/core/models/proprietary/dataset_generation/requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.2

# GitHub integration
PyGithub>=1.55
gitpython>=3.1.24

# Web interface
streamlit>=1.10.0

# Environment variables
python-dotenv>=0.19.0

# Code analysis
radon>=5.1.0
pylint>=2.12.0

# Testing
pytest>=6.2.5
pytest-cov>=2.12.1

# Development
black>=21.9b0
isort>=5.9.3
flake8>=3.9.2

# Voice transcription enhancements
SpeechRecognition>=3.8.1
pydub>=0.25.1
openai-whisper>=20231117
torch>=2.0.0
torchaudio>=2.0.0
numpy>=1.21.0
soundfile>=0.10.3 